requirements.txt
click>=8.0.0
jinja2>=3.1.0
pyyaml>=6.0
python-dotenv>=1.0.0
mcp>=1.0.0
rich>=13.0.0
pydantic>=2.0.0
typing-extensions>=4.0.0